// src/app/core/models/car.ts
export interface Car {
    brand: string;
    model: string;
    // km: number;
    km: string;
    // year: number;
    year: string;
    shift: string;
    fuel: string;
    color: string;
    car_type: string;
    e_car: boolean;
//     brand
// model of brand
// km
// year of the car
// shift type
// fuel type
// color of the car
// car type (sedan, suv, etc..)
// e-car
}